Move Fetch request to before index.html render
Move the Fetch request from main content rendering to before the
rendering of the entire page. This means that content is loaded before
the page, so there is not such a jarring effect on the loading of the
main section. It also enabled a rewrite to an async function (prompted
by vscode). The idea that this feeds into is that of a React project
structure, where all data flows downwards, with a central data point.
In the case of index.html, script.js will be this central point.